Organize workflow stream scenarios into directories
