+++ src/WorkflowStreams/BasicPublishSubscribe/Models.cs
@@ -1,0 +1,9 @@
+namespace TemporalioSamples.WorkflowStreams;
+
+using Temporalio.Extensions.WorkflowStreams;
+
+public record OrderInput(string OrderId, WorkflowStreamState? StreamState = null);
+
+public record StatusEvent(string Kind, string OrderId);
+
+public record ProgressEvent(string Message);
