+++ src/WorkflowStreams/ExternalPublisher/Models.cs
@@ -1,0 +1,7 @@
+namespace TemporalioSamples.WorkflowStreams;
+
+using Temporalio.Extensions.WorkflowStreams;
+
+public record HubInput(string HubId, WorkflowStreamState? StreamState = null);
+
+public record NewsEvent(string Headline);
