+++ src/WorkflowStreams/ReconnectingSubscriber/Models.cs
@@ -1,0 +1,10 @@
+namespace TemporalioSamples.WorkflowStreams;
+
+using Temporalio.Extensions.WorkflowStreams;
+
+public record PipelineInput(
+    string PipelineId,
+    TimeSpan? StageInterval = null,
+    WorkflowStreamState? StreamState = null);
+
+public record StageEvent(string Stage);
